Add serial support
