+++ go.mod
@@ -13,4 +13,7 @@ require (
 	github.com/sirupsen/logrus v1.6.0
 	github.com/stretchr/testify v1.4.0
 	github.com/x-cray/logrus-prefixed-formatter v0.5.2
+	go.bug.st/serial v1.1.0
+	golang.org/x/sys v0.0.0-20200806125547-5acd03effb82 // indirect
+	google.golang.org/appengine v1.6.6
 )
