+++ rest/mock.go
@@ -1,0 +1,14 @@
+package rest
+
+import (
+	"github.com/disaster37/go-arest"
+	"github.com/jarcoal/httpmock"
+)
+
+func MockRestClient() arest.Arest {
+	client := NewClient("http://localhost").(*Client)
+	httpmock.ActivateNonDefault(client.Client().GetClient())
+
+	return client
+
+}
